:sparkles: Use pydantic models instead of dataclasses (#GI-70)
