+++ dodo_is_api/models/channel_stop_types.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+__all__ = ('ChannelStopType',)
+
+
+class ChannelStopType(StrEnum):
+    COMPLETE = 'Complete'
+    REDIRECTION = 'Redirection'
