+++ dodo_is_api/models/delivery_transport_names.py
@@ -1,0 +1,9 @@
+from enum import StrEnum
+
+__all__ = ('DeliveryTransportName',)
+
+
+class DeliveryTransportName(StrEnum):
+    VEHICLE = 'Vehicle'
+    ON_FOOT = 'OnFoot'
+    BICYCLE = 'Bicycle'
