+++ dodo_is_api/models/sales_channels.py
@@ -1,0 +1,9 @@
+from enum import StrEnum
+
+__all__ = ('SalesChannel',)
+
+
+class SalesChannel(StrEnum):
+    DINE_IN = 'Dine-in'
+    TAKEAWAY = 'Takeaway'
+    DELIVERY = 'Delivery'
